Worked on player, projectile, enemy scaling

+++ Assets/Objects/Enemy/Enemy Health SO.cs
@@ -6,7 +6,7 @@ public class EnemyHealthSO : ScriptableObject
     public int maxHealth;
     
     public GameObject explosionPrefab;
-    public int explosionDuration;
+    public float explosionDuration;
     public float explosionFrequency;
     
     public GameObject reactorExplosionPrefab;
